Ship InMemoryConfigProvider #1713
