--- test/ReverseProxy.Tests/Common/InMemoryConfigProvider.cs
@@ -1,65 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using System.Collections.Generic;
-using System.Threading;
-using Microsoft.Extensions.Primitives;
-using Yarp.ReverseProxy.Configuration;
-using Yarp.ReverseProxy.Configuration.ConfigProvider;
-
-namespace Microsoft.Extensions.DependencyInjection
-{
-    public static class InMemoryConfigProviderExtensions
-    {
-        public static IReverseProxyBuilder LoadFromMemory(this IReverseProxyBuilder builder, IReadOnlyList<RouteConfig> routes, IReadOnlyList<ClusterConfig> clusters)
-        {
-            builder.Services.AddSingleton<IProxyConfigProvider>(new InMemoryConfigProvider(routes, clusters));
-            return builder;
-        }
-    }
-}
-
-namespace Yarp.ReverseProxy.Configuration.ConfigProvider
-{
-    public class InMemoryConfigProvider : IProxyConfigProvider
-    {
-        private volatile InMemoryConfig _config;
-
-        public InMemoryConfigProvider(IReadOnlyList<RouteConfig> routes, IReadOnlyList<ClusterConfig> clusters)
-        {
-            _config = new InMemoryConfig(routes, clusters);
-        }
-
-        public IProxyConfig GetConfig() => _config;
-
-        public void Update(IReadOnlyList<RouteConfig> routes, IReadOnlyList<ClusterConfig> clusters)
-        {
-            var newConfig = new InMemoryConfig(routes, clusters);
-            var oldConfig = Interlocked.Exchange(ref _config, newConfig);
-            oldConfig.SignalChange();
-        }
-
-        private class InMemoryConfig : IProxyConfig
-        {
-            private readonly CancellationTokenSource _cts = new CancellationTokenSource();
-
-            public InMemoryConfig(IReadOnlyList<RouteConfig> routes, IReadOnlyList<ClusterConfig> clusters)
-            {
-                Routes = routes;
-                Clusters = clusters;
-                ChangeToken = new CancellationChangeToken(_cts.Token);
-            }
-
-            public IReadOnlyList<RouteConfig> Routes { get; }
-
-            public IReadOnlyList<ClusterConfig> Clusters { get; }
-
-            public IChangeToken ChangeToken { get; }
-
-            internal void SignalChange()
-            {
-                _cts.Cancel();
-            }
-        }
-    }
-}
